Rewrite the sqlite create script parser.
Originally, we relied on CodeMirror's Lezer to build the AST and parsed from it. Now, we use our own tokenization approach. It produces a flat structure, and we build the parser on top of that. This gives us more flexibility for future extensions and reduces our dependency on external libraries.